requirements.txt
Flask==1.1.2
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
itsdangerous==1.1.0
MarkupSafe==1.1.1
numpy>=1.9.2
scipy>=0.15.1
scikit-learn==0.23.1
matplotlib>=1.4.3
pandas>=0.19
nltk>=3.4.5
joblib==0.15.1
mysql-connector
flask-cors
